final version

+++ ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: lbarreta <lbarreta@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/26 17:47:36 by lbarreta          #+#    #+#             */
-/*   Updated: 2020/02/19 22:10:25 by lbarreta         ###   ########.fr       */
+/*   Updated: 2020/02/21 02:36:10 by lbarreta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, size_t size)
+size_t		ft_strlcpy(char *dest, char *src, size_t size)
 {
 	size_t i;
